+++ packages/web/components/Technology/TechnologyProvider.js
@@ -1,10 +1,23 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import TechnologyContext from './TechnologyContext';
+import { formatMoney } from '../../utils/helper';
 
-export const TechnologyProvider = ({ children, technology }) => (
-	<TechnologyContext.Provider value={{ technology }}>{children}</TechnologyContext.Provider>
-);
+export const TechnologyProvider = ({ children, technology }) => {
+	const implementationCosts = useMemo(() => {
+		const costs = technology?.technologyCosts?.costs?.implementation_costs;
+
+		const total = costs?.reduce((acc, item) => acc + item?.quantity * item?.value, 0);
+
+		return formatMoney(total);
+	}, [technology]);
+
+	return (
+		<TechnologyContext.Provider value={{ technology, implementationCosts }}>
+			{children}
+		</TechnologyContext.Provider>
+	);
+};
 
 TechnologyProvider.propTypes = {
 	children: PropTypes.node.isRequired,
